requirements.txt
EasyProcess==0.3
entrypoint2==0.2.3
evdev==1.4.0
jeepney==0.6.0
mss==6.1.0
Pillow==8.1.2
pynput==1.7.3
pyscreenshot==2.3
python-xlib==0.29
six==1.15.0

numpy~=1.20.1
future~=0.18.2